--- templates/index.html
@@ -1,49 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="{{.rootPath}}/public/styles.css">
-    <title>TM-to-ABRP</title>
-</head>
-
-<body>
-    <header>
-        <h1>TM-to-ABRP</h1>
-    </header>
-
-    <main>
-        <h2>Car</h2>
-        <p>State: {{.carState}} (Previous: {{.carPreviousState}})</p>
-
-        <h5>ABRP Updates</h5>
-        <table>
-            <tr>
-                <td>Now?</td>
-                <td>
-                    <form action="{{.rootPath}}/abrp-send-now" method="post">
-                        <button type="submit">Go</button>
-                    </form>
-                </td>
-            </tr>
-            <tr>
-                <td>Continuous?</td>
-                <td>
-                    <form action="{{.rootPath}}/abrp-send-continuous" method="post">
-                        <input type="datetime-local" name="endtime" value="{{.carAbrpUpdatesEndTimeString}}" />
-                        <button type="submit">{{.carAbrpSendContinuousButtonText}}</button>
-                    </form>
-                </td>
-            </tr>
-        </table>
-
-        <h5>TM Data</h5>
-        <pre>{{.carTmData}}</pre>
-
-        <h5>ABRP Data</h5>
-        <pre>{{.carAbrpData}}</pre>
-    </main>
-</body>
-
-</html>
